# openapi_base_gabinete_ia.yaml
openapi: 3.0.3
info:
  title: GABINETE IA API
  version: 1.0.0
  description: >-
    Contrato base OpenAPI/Swagger do sistema GESTÃO DE MANDATOS / GABINETE IA.
    API REST versionada para operação web e mobile.
servers:
  - url: /api/v1
    description: Base URL v1
...
security:
  - bearerAuth: []
paths:
  /auth/login:
    post:
      tags: [Auth]
      summary: Autenticar usuário
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login efetuado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /auth/me:
    get:
      tags: [Auth]
      summary: Retornar contexto do usuário autenticado
      responses:
        '200':
          description: Contexto do usuário
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /contatos:
    get:
      tags: [Contatos]
      summary: Listar contatos/cidadãos
      parameters:
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: territorio_id
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/StatusContato'
        - name: tipo_contato
          in: query
          schema:
            type: string
        - name: duplicidade_suspeita
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Lista de contatos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactListResponse'
    post:
      tags: [Contatos]
      summary: Criar contato/cidadão
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContactCreateRequest'
      responses:
        '201':
          description: Contato criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactSingleResponse'
        '409':
          $ref: '#/components/responses/ConflictError'
...
  /demandas:
    get:
      tags: [Demandas]
      summary: Listar demandas
      parameters:
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/DateFrom'
        - $ref: '#/components/parameters/DateTo'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/StatusDemanda'
        - name: prioridade
          in: query
          schema:
            $ref: '#/components/schemas/PrioridadeDemanda'
        - name: categoria_id
          in: query
          schema:
            type: string
            format: uuid
        - name: territorio_id
          in: query
          schema:
            type: string
            format: uuid
        - name: responsavel_usuario_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Lista paginada de demandas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DemandListResponse'
    post:
      tags: [Demandas]
      summary: Criar demanda
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DemandCreateRequest'
      responses:
        '201':
          description: Demanda criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DemandSingleResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
...
  /demandas/{demanda_id}/historico:
    parameters:
      - $ref: '#/components/parameters/DemandaId'
    get:
      tags: [Demandas]
      summary: Listar histórico da demanda
      responses:
        '200':
          description: Histórico da demanda
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DemandHistoryListResponse'
...
  /oficios:
    get:
      tags: [Jurídico]
      summary: Listar ofícios
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Lista de ofícios
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LegalDocumentListResponse'
    post:
      tags: [Jurídico]
      summary: Criar ofício
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReqOficioCreateRequest'
      responses:
        '201':
          description: Ofício criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LegalDocumentSingleResponse'
...
  /territorial/dashboard:
    get:
      tags: [Territorial]
      summary: Dashboard territorial agregado
      parameters:
        - $ref: '#/components/parameters/DateFrom'
        - $ref: '#/components/parameters/DateTo'
        - name: regiao_id
          in: query
          schema:
            type: string
            format: uuid
        - name: bairro_id
          in: query
          schema:
            type: string
            format: uuid
        - name: microarea_id
          in: query
          schema:
            type: string
            format: uuid
        - name: tema
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Dashboard territorial agregado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TerritorialDashboardResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Search:
      name: search
      in: query
      schema:
        type: string
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    DateFrom:
      name: date_from
      in: query
      schema:
        type: string
        format: date-time
    DateTo:
      name: date_to
      in: query
      schema:
        type: string
        format: date-time
...
    DemandaId:
      name: demanda_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    UnauthorizedError:
      description: Não autenticado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ValidationError:
      description: Erro de validação
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ConflictError:
      description: Conflito de recurso ou duplicidade
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    PerfilUsuario:
      type: string
      enum:
        - COLABORADOR_EXTERNO
        - SUPERVISOR_EQUIPE
        - ASSESSOR_NIVEL_1
        - ASSESSOR_JURIDICO
        - ASSESSOR_ADMINISTRATIVO
        - CHEFE_GABINETE
        - VEREADOR
    StatusContato:
      type: string
      enum: [ATIVO, INATIVO, PENDENTE_VALIDACAO, DUPLICIDADE_SUSPEITA]
    StatusDemanda:
      type: string
      enum: [ABERTA, EM_TRIAGEM, EM_ATENDIMENTO, ENCAMINHADA, AGUARDANDO_RETORNO, CONCLUIDA, SUSPENSA, CANCELADA, REABERTA]
    PrioridadeDemanda:
      type: string
      enum: [BAIXA, MEDIA, ALTA, CRITICA]
...
    StatusDocumento:
      type: string
      enum: [RASCUNHO, EM_ELABORACAO, EM_REVISAO, APROVADO, PROTOCOLADO, ARQUIVADO, CANCELADO]
    TipoDocumento:
      type: string
      enum: [PARECER, REQUERIMENTO, OFICIO, MINUTA, ANEXO_OPERACIONAL, OUTRO]
...
    OrigemCadastro:
      type: string
      enum: [MOBILE_CAMPO, WEB_INTERNO, IMPORTACAO, INTEGRACAO]
...
    Meta:
      type: object
      required: [request_id, timestamp]
      properties:
        request_id:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
    PaginationMeta:
      allOf:
        - $ref: '#/components/schemas/Meta'
        - type: object
          required: [page, page_size, total, total_pages]
          properties:
            page:
              type: integer
            page_size:
              type: integer
            total:
              type: integer
            total_pages:
              type: integer
    ErrorDetail:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
    ErrorObject:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'
    ErrorResponse:
      type: object
      required: [error, meta]
      properties:
        error:
          $ref: '#/components/schemas/ErrorObject'
        meta:
          $ref: '#/components/schemas/Meta'
    ScopeInput:
      type: object
      required: [escopo_tipo]
      properties:
        escopo_tipo:
          type: string
        territorio_id:
          type: string
          format: uuid
        equipe_id:
          type: string
          format: uuid
        escopo_valor:
          type: string
    ScopeOutput:
      allOf:
        - $ref: '#/components/schemas/ScopeInput'
        - type: object
          properties:
            id:
              type: string
              format: uuid
    LoginRequest:
      type: object
      required: [email_login, senha]
      properties:
        email_login:
          type: string
          format: email
        senha:
          type: string
          format: password
...
    TokenBundle:
      type: object
      required: [access_token, refresh_token, token_type, expires_in]
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          example: bearer
        expires_in:
          type: integer
    UserSummary:
      type: object
      required: [id, nome, perfil, gabinete_id]
      properties:
        id:
          type: string
          format: uuid
        nome:
          type: string
        perfil:
          $ref: '#/components/schemas/PerfilUsuario'
        gabinete_id:
          type: string
          format: uuid
        equipe_id:
          type: string
          format: uuid
          nullable: true
        ativo:
          type: boolean
        escopos:
          type: array
          items:
            $ref: '#/components/schemas/ScopeOutput'
    LoginResponse:
      type: object
      properties:
        data:
          allOf:
            - $ref: '#/components/schemas/TokenBundle'
            - type: object
              properties:
                user:
                  $ref: '#/components/schemas/UserSummary'
        meta:
          $ref: '#/components/schemas/Meta'
...
    MeResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/UserSummary'
        meta:
          $ref: '#/components/schemas/Meta'
...
    ContactBase:
      type: object
      properties:
        territorio_id:
          type: string
          format: uuid
          nullable: true
        origem_cadastro:
          $ref: '#/components/schemas/OrigemCadastro'
        nome:
          type: string
        cpf:
          type: string
          nullable: true
        data_nascimento:
          type: string
          format: date
          nullable: true
        telefone_principal:
          type: string
          nullable: true
        telefone_secundario:
          type: string
          nullable: true
        email:
          type: string
          format: email
          nullable: true
        logradouro:
          type: string
          nullable: true
        numero:
          type: string
          nullable: true
        complemento:
          type: string
          nullable: true
        bairro:
          type: string
          nullable: true
        cidade:
          type: string
          nullable: true
        cep:
          type: string
          nullable: true
        tipo_contato:
          type: string
        observacoes:
          type: string
          nullable: true
    ContactCreateRequest:
      allOf:
        - $ref: '#/components/schemas/ContactBase'
        - type: object
          required: [origem_cadastro, nome]
...
    ContactItem:
      allOf:
        - $ref: '#/components/schemas/ContactBase'
        - type: object
          properties:
            id:
              type: string
              format: uuid
            status:
              $ref: '#/components/schemas/StatusContato'
            duplicidade_suspeita:
              type: boolean
            territorio_nome:
              type: string
              nullable: true
            created_at:
              type: string
              format: date-time
...
    ContactSingleResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/ContactItem'
        meta:
          $ref: '#/components/schemas/Meta'
...
    ContactListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ContactItem'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    DemandBase:
      type: object
      properties:
        cidadao_id:
          type: string
          format: uuid
          nullable: true
        territorio_id:
          type: string
          format: uuid
          nullable: true
        categoria_id:
          type: string
          format: uuid
          nullable: true
        titulo:
          type: string
        descricao:
          type: string
        prioridade:
          $ref: '#/components/schemas/PrioridadeDemanda'
        origem_cadastro:
          $ref: '#/components/schemas/OrigemCadastro'
    DemandCreateRequest:
      allOf:
        - $ref: '#/components/schemas/DemandBase'
        - type: object
          required: [titulo, descricao, prioridade, origem_cadastro]
...
    DemandItem:
      allOf:
        - $ref: '#/components/schemas/DemandBase'
        - type: object
          properties:
            id:
              type: string
              format: uuid
            status:
              $ref: '#/components/schemas/StatusDemanda'
            responsavel_usuario_id:
              type: string
              format: uuid
              nullable: true
            responsavel_nome:
              type: string
              nullable: true
            categoria_nome:
              type: string
              nullable: true
            territorio_nome:
              type: string
              nullable: true
            cidadao_nome:
              type: string
              nullable: true
            sla_data:
              type: string
              format: date-time
              nullable: true
            criticidade_derivada:
              type: string
              nullable: true
            data_abertura:
              type: string
              format: date-time
...
    DemandSingleResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/DemandItem'
        meta:
          $ref: '#/components/schemas/Meta'
...
    DemandListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/DemandItem'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    DemandHistoryItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        acao:
          type: string
        status_anterior:
          $ref: '#/components/schemas/StatusDemanda'
        status_novo:
          $ref: '#/components/schemas/StatusDemanda'
        observacao:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        usuario:
          type: object
          properties:
            id:
              type: string
              format: uuid
            nome:
              type: string
    DemandHistoryListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/DemandHistoryItem'
        meta:
          $ref: '#/components/schemas/Meta'
...
    LegalDocumentBase:
      type: object
      properties:
        demanda_id:
          type: string
          format: uuid
          nullable: true
        protocolo_id:
          type: string
          format: uuid
          nullable: true
        projeto_id:
          type: string
          format: uuid
          nullable: true
        titulo:
          type: string
        status:
          $ref: '#/components/schemas/StatusDocumento'
        conteudo_texto:
          type: string
          nullable: true
...
    ReqOficioCreateRequest:
      allOf:
        - $ref: '#/components/schemas/LegalDocumentBase'
        - type: object
          required: [titulo, status]
          properties:
            assunto:
              type: string
              nullable: true
            destinatario:
              type: string
              nullable: true
...
    LegalDocumentItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tipo:
          $ref: '#/components/schemas/TipoDocumento'
        titulo:
          type: string
        status:
          $ref: '#/components/schemas/StatusDocumento'
        versao_atual:
          type: integer
        autor_usuario_id:
          type: string
          format: uuid
          nullable: true
        demanda_id:
          type: string
          format: uuid
          nullable: true
    LegalDocumentSingleResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/LegalDocumentItem'
        meta:
          $ref: '#/components/schemas/Meta'
...
    LegalDocumentListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/LegalDocumentItem'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    TerritorialDashboardResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            cards:
              type: object
              properties:
                quantidade_demandas:
                  type: integer
                quantidade_visitas:
                  type: integer
                quantidade_eventos:
                  type: integer
                tempo_medio_atendimento_horas:
                  type: number
                  nullable: true
            mapa:
              type: array
              items:
                type: object
                properties:
                  territorio_id:
                    type: string
                    format: uuid
                  territorio_nome:
                    type: string
                  quantidade_demandas:
                    type: integer
            temas:
              type: array
              items:
                type: object
                properties:
                  tema:
                    type: string
                  quantidade:
                    type: integer
        meta:
          $ref: '#/components/schemas/Meta'
...